Fixed storyboard.
Summary:
I enabled join and subscribe buttons to add members or subscribers to NSArrays. I also showed new buttons if you are a member of the group.

I added basics of events and how an event will look.

I added more details to FBUEventDetailViewController, but it is not completely finished.

Reviewers: christineyen

Reviewed By: christineyen

CC: umag, ameighan, nlutsenko

Differential Revision: https://phab.parse.com/D5426